Fix for bug 161
Description:
  This is only an initial draft of the fix and is not the complete fix.
  This is only for review purpose to resolve a build issue due to the code.

+++ opendaylight/protocol_plugins/openflow/src/main/java/org/opendaylight/controller/protocol_plugin/openflow/internal/DataPacketMuxDemux.java
@@ -296,7 +296,7 @@ public class DataPacketMuxDemux implements IContainerListener,
             return;
         }
 
-        Short port = (Short) outPort.getID();
+        int port = (int) outPort.getID();
         Long swID = (Long) outPort.getNode().getID();
         ISwitch sw = this.swID2ISwitch.get(swID);
